Hokey first pass to enable SMAC to prevent stray accesses to user memory.
To enable SMAC under qemu pass something like: -cpu Haswell,+smac

We need to gate this on CPUID since it is a new feature and processors
may fault if they don't support it.

Also the kernel doesn't work at all because there are a ton of places
that access user memory. Most of them *should* be using
copy_{to,from}_user, though, and should be an easy fix (and a good way
to find them all!).